requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=21.0
